Update to get stats starting to work
Need to look at permissions.  The stats were not working because the motion user did not have access to the location of the .dat file to update daily events.  Solved in a crude way by modifying the permissions of the file, but a better way would be to locate the statistics files in a common area.